+++ AngouriMath/Functions/Algebra/EquationSolver/Numerical/NewtonSolver.cs
@@ -55,7 +55,7 @@ namespace AngouriMath.Functions.Algebra.NumbericalSolving
                 if (i > minCheckIters && prev == value)
                     return value;
             }
-            if (Number.Abs(prev - value) > MathS.Settings.PrecisionError) // TODO: threshold
+            if (Number.Abs(prev - value) > MathS.Settings.PrecisionErrorCommon)
                 return RealNumber.NaN();
             else
                 return value;
@@ -85,8 +85,7 @@ namespace AngouriMath.Functions.Algebra.NumbericalSolving
             (decimal Re, decimal Im) to, 
             (int Re, int Im) stepCount, int precision)
         {
-            var userIterCount = MathS.Settings.FloatToRationalIterCount;
-            MathS.Settings.FloatToRationalIterCount = -1;
+            MathS.Settings.FloatToRationalIterCount.Set(0);
             var res = new Set();
             var df = expr.Derive(v).Simplify().Compile(v);
             var f = expr.Simplify().Compile(v);
@@ -98,10 +97,10 @@ namespace AngouriMath.Functions.Algebra.NumbericalSolving
                     var value = Number.Create(from.Re * xShare + to.Re * (1 - xShare),
                                            from.Im * yShare + to.Im * (1 - yShare));
                     var root = NewtonIter(f, df, value, precision);
-                    if (root.IsDefinite() && f.Call(root).Abs() < MathS.Settings.PrecisionError)
+                    if (root.IsDefinite() && f.Call(root).Abs() < MathS.Settings.PrecisionErrorCommon)
                         res.Add(root);
                 }
-            MathS.Settings.FloatToRationalIterCount = userIterCount;
+            MathS.Settings.FloatToRationalIterCount.Unset();
             return res;
         }
     }